[compiler] Todo for fbt with multiple pronoun/plural
[ghstack-poisoned]
